Se configura la inyección de dependencias con el UserRepository para poder consumir el servicio de crear cuenta
